Adding DangerContext as part of the receiver context in both danger() and rule() blocks
